// src/app/interfaces/common/nineteen-points.interface.ts
export interface NineteenPoints {
  _id?: string;
  name?: string;
  image?: string;
  seoImage?: string;
  images?: string[];
  informations?: any[];
  description?: string;
  isHtml?: boolean;
  descriptionEn?: string;
  isHtmlEn?: boolean;
  websiteUrl?: string;
  slug?: string;
  createdAt?: Date;
  updatedAt?: Date;
  nameEn?: string;
}